Some small fixes to avoid warning messages from g++
git-svn-id: svn://irodssvn.ucsd.edu/trunk@3953 ea137bd6-aa86-47b6-b90d-b101d9f19a90

+++ iRODS/server/icat/src/icatHighLevelRoutines.c
@@ -2949,7 +2949,7 @@ int chlSimpleQuery(rsComm_t *rsComm, char *sql,
       }
       rstrcat(outBuf, "\n", maxOutBuf);
       if (rowSize==0) rowSize=strlen(outBuf);
-      if (strlen(outBuf)+rowSize+20 > maxOutBuf) {
+      if ((int)strlen(outBuf)+rowSize+20 > maxOutBuf) {
 	 return(0); /* success so far, but more rows available */
       }
    }
